+++ fms_accel/scattermoe_utils/__init__.py
@@ -1,0 +1,4 @@
+from .expert_parallel import (
+    all_to_all_gather_inputs,
+    scatter_with_routing_weights
+)
